requirements.txt
dj-database-url
Django
psycopg2
python-decouple
gunicorn
whitenoise
validate_email